Moving resources to reconcilers
